Replace photos with better-compressed versions, add logout button on more pages
