+++ index.html
@@ -3,7 +3,7 @@
 
 <head>
   <meta charset="UTF-8" />
-  <link rel="icon" type="image/png" href="/QueryFlow-icon.png" />
+  <link rel="icon" type="image/avif" href="/QueryFlow-icon.avif" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>QueryFlow</title>
 </head>
