+++ src/Pages/UserLogin.svelte
@@ -1,0 +1,41 @@
+<script lang="ts">
+	import Login from '../lib/Login.svelte';
+	import Signup from '../lib/Signup.svelte';
+	import graph from '../assets/graph3.avif';
+	import { navigate } from 'svelte-routing';
+	import { onMount } from 'svelte';
+	// toggle signup/login component
+	let renderSignup: boolean = false;
+
+	onMount(async () => {
+		if (localStorage.getItem('token')) navigate('/home', { replace: true });
+	});
+</script>
+
+<div class="landing-login-container flex flex-col mt-5 lg:flex-row">
+	<div
+		class="flex-1 flex flex-col items-center justify-center px-6 py-8 mx-auto lg:py-0 responsive-column"
+	>
+		<img class="object-contain rounded-lg" src={graph} alt="graph" />
+	</div>
+	<!-- conditional statement to render signup or login -->
+	{#if renderSignup}
+		<Signup bind:renderSignup />
+	{:else}
+		<Login bind:renderSignup />
+	{/if}
+</div>
+
+<style>
+
+	.landing-login-container {
+		height: 80vh;
+	}
+	.responsive-column {
+		height: 90vh;
+	}
+
+	@media only screen and (max-width: 1025px) {
+		
+	}
+</style>
